// libs/ui/src/lib/hint/types.ts
export type HintDirection =
  | 'bottom-left'
  | 'bottom-right'
  | 'bottom'
  | 'left-bottom'
  | 'left-top'
  | 'left'
  | 'right-bottom'
  | 'right-top'
  | 'right'
  | 'top-left'
  | 'top-right'
  | 'top';

export const HINT_DIRECTIONS: HintDirection[] = [
  'bottom-left',
  'bottom-right',
  'bottom',
  'left-bottom',
  'left-top',
  'left',
  'right-bottom',
  'right-top',
  'right',
  'top-left',
  'top-right',
  'top',
];